+++ plop-templates/component/ClientComponent.tsx.hbs
@@ -1,5 +1,3 @@
-'use client';
-
 import { FC } from 'react';
 
 import style from './index.module.scss';
