--- src/app/(account)/before/page.tsx
@@ -1,9 +1,0 @@
-import { NextPage } from 'next';
-
-import { SignUpLoginBefore } from '@/components/page/signup/SignUpLoginBefore';
-
-const SignUpLoginBeforePage: NextPage = () => {
-  return <SignUpLoginBefore />;
-};
-
-export default SignUpLoginBeforePage;
