--- src/app/(account)/login/page.tsx
@@ -1,10 +1,0 @@
-import { NextPage } from 'next';
-import React from 'react';
-
-import { Login } from '@/components/page/Login';
-
-const LoginPage: NextPage = () => {
-  return <Login />;
-};
-
-export default LoginPage;
