--- src/app/(account)/signup/page.tsx
@@ -1,10 +1,0 @@
-import { NextPage } from 'next';
-import React from 'react';
-
-import { SignUp } from '@/components/page/signup/SignUp';
-
-const SignUpPage: NextPage = () => {
-  return <SignUp />;
-};
-
-export default SignUpPage;
