--- src/app/(loggingIn)/(hasLayoutHeader)/dish/[dishId]/page.tsx
@@ -1,9 +1,0 @@
-import { NextPage } from 'next';
-
-import { Dish } from '@/components/page/Dish';
-
-const DishPage: NextPage = () => {
-  return <Dish />;
-};
-
-export default DishPage;
