--- src/app/(loggingIn)/(hasLayoutHeader)/user/page.tsx
@@ -1,9 +1,0 @@
-import { NextPage } from 'next';
-
-import { User } from '@/components/page/User';
-
-const UserPage: NextPage = () => {
-  return <User />;
-};
-
-export default UserPage;
