--- src/app/(loggingIn)/page.tsx
@@ -1,9 +1,0 @@
-import type { NextPage } from 'next';
-
-import { Home } from '@/components/page/Home';
-
-const HomePage: NextPage = () => {
-  return <Home />;
-};
-
-export default HomePage;
