--- src/app/_app.tsx
@@ -1,8 +1,0 @@
-import type { AppProps } from 'next/app';
-import React from 'react';
-
-function MyApp({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />;
-}
-
-export default MyApp;
