+++ src/assets/styles/globals.css
@@ -26,6 +26,12 @@ input {
   outline: 0;
 }
 
+/* stylelint-disable-next-line selector-id-pattern */
+#__next {
+  width: 100%;
+  height: 100%;
+}
+
 /** ダークモードか判定してcssを変更できるが、今回はしない */
 
 /* @media (prefers-color-scheme: dark) {
