+++ src/components/layout/default/index.module.scss
@@ -1,4 +1,4 @@
-@import '../../assets/styles/resources';
+@import '../../../assets/styles/resources';
 
 .home-layout {
   position: relative;
