+++ src/components/layout/default/index.ts
@@ -1,0 +1,1 @@
+export * from './defaultLayout';
