+++ src/components/layout/hasHeader/index.ts
@@ -1,0 +1,1 @@
+export * from './hasHeaderLayout';
