+++ src/components/model/dish/DishItem/DishItem.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import { FC } from 'react';
 
 import { DishTime } from '../DishTime';
