+++ src/components/page/Dish/Dish.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import { useParams } from 'next/navigation';
 import { FC } from 'react';
