+++ src/components/page/User/User.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import { useSearchParams } from 'next/navigation';
 import { FC } from 'react';
