+++ src/components/page/signup/SignUpLoginBefore/SignUpLoginBefore.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import { useRouter } from 'next/navigation';
 import { FC } from 'react';
