+++ src/components/section/dish/DishDetail/DishDetail.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import { FC } from 'react';
 import { Slide } from 'react-slideshow-image';
 import 'react-slideshow-image/dist/styles.css';
