+++ src/components/section/signup/SignUpCode/SignUpCode.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import { useRouter } from 'next/navigation';
 import { FC, useContext } from 'react';
