+++ src/components/section/signup/SignUpConfirm/SignUpConfirm.tsx
@@ -1,4 +1,3 @@
-'use client';
 import { useRouter } from 'next/navigation';
 import { FC, useEffect } from 'react';
