+++ src/components/section/signup/SignUpField/SignUpField.tsx
@@ -1,4 +1,3 @@
-'use client';
 import { useRouter } from 'next/navigation';
 import { ChangeEvent, FC } from 'react';
