+++ src/components/section/user/UserDetail/UserDetail.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import clsx from 'clsx';
 import { useRouter } from 'next/navigation';
 import { FC } from 'react';
