+++ src/components/ui/Button/Button.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import { clsx } from 'clsx';
 import { FC, MouseEventHandler } from 'react';
