+++ src/components/ui/Header/Header.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import clsx from 'clsx';
 import { useRouter, usePathname } from 'next/navigation';
 import { destroyCookie } from 'nookies';
