+++ src/components/ui/filter/FilterPanel/FilterPanel.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import { FC, useEffect, useRef, useState } from 'react';
 import { CSSTransition } from 'react-transition-group';
