+++ src/components/ui/form/FormCheckbox/FormCheckbox.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import clsx from 'clsx';
 import { FC } from 'react';
