+++ src/components/ui/form/FormResult/FormResult.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import { FC } from 'react';
 
 import { FormTitle } from '../FormTitle';
