+++ src/components/ui/form/FormSearch/FormSearch.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import clsx from 'clsx';
 import { ChangeEventHandler, FC, useRef } from 'react';
