+++ src/components/ui/form/FormTitle/FormTitle.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import { FC } from 'react';
 
 import RequiredIcon from '@/assets/icons/required.svg';
